Added permission checking in the backend

+++ server/src/controllers/org_environment_variables.controller.ts
@@ -1,17 +1,30 @@
-import { Controller, Post, UseGuards, Body, Get, Patch, Delete, Param, BadRequestException } from '@nestjs/common';
+import {
+  Controller,
+  Post,
+  UseGuards,
+  Body,
+  Get,
+  Patch,
+  Delete,
+  Param,
+  BadRequestException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { decamelizeKeys } from 'humps';
 import { JwtAuthGuard } from '../modules/auth/jwt-auth.guard';
-import { AppAbility } from 'src/modules/casl/casl-ability.factory';
 import { PoliciesGuard } from 'src/modules/casl/policies.guard';
-import { CheckPolicies } from 'src/modules/casl/check_policies.decorator';
-import { User as UserEntity } from 'src/entities/user.entity';
 import { User } from 'src/decorators/user.decorator';
 import { EnvironmentVariableDto } from '@dto/environment-variable.dto';
 import { OrgEnvironmentVariablesService } from '@services/org_environment_variables.service';
+import { OrgEnvironmentVariablesAbilityFactory } from 'src/modules/casl/abilities/org-environment-variables-ability.factory';
+import { OrgEnvironmentVariable } from 'src/entities/org_envirnoment_variable.entity';
 
 @Controller('organization-variables')
 export class OrgEnvironmentVariablesController {
-  constructor(private orgEnvironmentVariablesService: OrgEnvironmentVariablesService) {}
+  constructor(
+    private orgEnvironmentVariablesService: OrgEnvironmentVariablesService,
+    private orgEnvironmentVariablesAbilityFactory: OrgEnvironmentVariablesAbilityFactory
+  ) {}
 
   @UseGuards(JwtAuthGuard)
   @Get()
@@ -22,25 +35,40 @@ export class OrgEnvironmentVariablesController {
 
   // Endpoint for adding new env vars
   @UseGuards(JwtAuthGuard, PoliciesGuard)
-  @CheckPolicies((ability: AppAbility) => ability.can('addOrgEnvironmentVariables', UserEntity))
   @Post()
   async create(@User() user, @Body() environmentVariableDto: EnvironmentVariableDto) {
+    const ability = await this.orgEnvironmentVariablesAbilityFactory.orgEnvironmentVariableActions(user, {});
+
+    if (!ability.can('createOrgEnvironmentVariable', OrgEnvironmentVariable)) {
+      throw new ForbiddenException('You do not have permissions to perform this action');
+    }
+
     const result = await this.orgEnvironmentVariablesService.create(user, environmentVariableDto);
     return decamelizeKeys({ variable: result });
   }
 
   @UseGuards(JwtAuthGuard, PoliciesGuard)
-  @CheckPolicies((ability: AppAbility) => ability.can('updateOrgEnvironmentVariables', UserEntity))
   @Patch(':id')
   async update(@Body() body, @User() user, @Param('id') variableId) {
+    const ability = await this.orgEnvironmentVariablesAbilityFactory.orgEnvironmentVariableActions(user, {});
+
+    if (!ability.can('updateOrgEnvironmentVariable', OrgEnvironmentVariable)) {
+      throw new ForbiddenException('You do not have permissions to perform this action');
+    }
+
     await this.orgEnvironmentVariablesService.update(user.organizationId, variableId, body);
     return {};
   }
 
   @UseGuards(JwtAuthGuard, PoliciesGuard)
-  @CheckPolicies((ability: AppAbility) => ability.can('deleteOrgEnvironmentVariables', UserEntity))
   @Delete(':id')
   async delete(@User() user, @Param('id') variableId) {
+    const ability = await this.orgEnvironmentVariablesAbilityFactory.orgEnvironmentVariableActions(user, {});
+
+    if (!ability.can('deleteOrgEnvironmentVariable', OrgEnvironmentVariable)) {
+      throw new ForbiddenException('You do not have permissions to perform this action');
+    }
+
     const result = await this.orgEnvironmentVariablesService.delete(user.organizationId, variableId);
     if (result.affected == 1) {
       return;
